requirements.txt
PyQtChart == 5.15.16
PyQt5-stubs == 5.15.6.0
scipy == 1.11.2
